Incorporate docs content into main website
Instead of redirecting to docs.bpftrace.org,
include the docs in this website.
This is so we don't have to sync css and
frame changes to the docs site. Also there
is no longer inconsistency (e.g. the versions
dropdown) between the two.

This also adds a node script (make-doc.js)
that converts an html file generated from asciidoctor
into a version page for the website
(details on use in the script).

Next step is to add a github action to run
asciidoctor and the make-doc.js script to
automatically sync changes from the bpftrace.adoc
in the main repo to the pre-release docs here.